+++ src/hooks/index.ts
@@ -1,0 +1,3 @@
+export { useInput } from '@/hooks/useInput';
+export { useTabs } from '@/hooks/useTabs';
+export { useTitle } from '@/hooks/useTitle';
